requirement.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.3.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
bce-python-sdk==0.8.74
beautifulsoup4==4.6.3
bitarray==2.6.0
black==22.12.0
bleach==6.0.0
blis==0.7.8
blobfile==2.0.0
boto3==1.26.16
botocore==1.29.16
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cityscapesScripts==2.2.1
click==8.1.3
cloudpickle==2.2.0
cmd2==2.4.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.2
confection==0.0.2
contextlib2==21.6.0
contourpy==1.0.5
crcmod==1.7
cryptography==38.0.4
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
dataclasses==0.6
datasets==2.5.2
debugpy==1.6.4
decorator==5.1.1
deepspeed==0.6.5
defusedxml==0.7.1
dgl==0.9.1
dgl-cu113==0.9.1
dglgo==0.0.2
dill==0.3.5.1
docker-pycreds==0.4.0
docstring-to-markdown==0.11
docutils==0.19
easydict==1.10
einops==0.5.0
embeddings==0.0.8
emoji==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
fairscale==0.4.12
-e git+ssh://git@github.com/yiting-tom/ICCV.git@60c7cb9780b71ca1556c5060c628bf41aa1c8381#egg=fairseq&subdirectory=fairseq
fastai==2.7.9
fastcore==1.5.27
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.8.2
fire==0.4.0
Flask==2.2.2
Flask-Babel==2.0.0
flatbuffers==22.9.24
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.5.3
gdown==4.6.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
hjson==3.1.0
html5lib==1.1
huggingface-hub==0.10.0
humanfriendly==10.0
hydra-core==1.0.7
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==5.1.0
importlib-resources==5.10.2
iniconfig==1.1.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.2
jedi-language-server==0.40.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json-tricks==3.15.5
jsonplus==0.8.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
kiwisolver==1.4.4
langcodes==3.3.0
language-evaluation @ git+https://github.com/j-min/language-evaluation@507417ec164e875b611d1ce5e0b8543cec77c429
lanms-neo==1.0.2
Levenshtein==0.20.8
lightgbm==3.3.3
lightning-utilities==0.4.2
littleutils==0.2.2
lmdb==1.3.0
lpips==0.1.4
lsprotocol==2022.0.0a10
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
-e git+https://github.com/microsoft/GLIP.git@91f20921e691922167d8fa94fb55fa7a25ffe0ff#egg=maskrcnn_benchmark
matplotlib==3.6.2
matplotlib-inline==0.1.6
maven==0.0.6
MechanicalSoup==0.8.0
megatron-util==1.1.0
mistune==2.0.5
ml-collections==0.1.1
mmcls==0.24.0
mmdet==2.25.2
mock==4.0.3
modelscope==1.1.4
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mypy==0.991
mypy-extensions==0.4.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.6
ninja==1.11.1
nltk==3.7
nmslib==2.1.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
numpydoc==1.4.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-htop==1.0.5
oauthlib==3.2.2
ofa==0.1.0.post202203231606
ogb==1.3.4
omegaconf==2.0.6
OnedriveCMD==0.1.8.1
onedrivesdk==2.0.1
onnxruntime==1.12.1
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
oss2==2.16.0
outdated==0.2.1
packaging==22.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.0
paddlepaddle==2.4.0
paddlepaddle-gpu==2.4.0
pai-easycv==0.6.3.6
pai-easynlp==0.0.7
pandarallel==1.6.4
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pathy==0.6.2
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
pluggy==1.0.0
Polygon3==3.0.9.1
portalocker==2.5.1
premailer==3.10.0
preshed==3.0.7
prettytable==0.7.2
progress==1.6
progressbar2==3.34.3
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pyclipper==1.3.0.post4
pycocoevalcap==1.2
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.16.0
pydantic==1.9.2
pydot==1.4.2
pygls==1.0.1
Pygments==2.13.0
pyjnius==1.4.2
PyMuPDF==1.19.0
pyparsing==3.0.9
pyperclip==1.8.2
pyquaternion==0.9.9
pyrsistent==0.19.3
PySide6-Essentials==6.4.2
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.6
python-Levenshtein==0.20.8
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-slugify==7.0.0
python-utils==3.4.5
pytorch-lightning==1.8.4.post0
pytorch-metric-learning==1.6.2
pytorch-pretrained-bert==0.6.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
rapidfuzz==2.13.2
rdkit-pypi==2022.3.5
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge==1.0.1
rouge-score==0.0.4
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sacrebleu==2.2.1
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.9.9
seqeval==1.2.2
setproctitle==1.3.2
sh==1.14.3
Shapely==1.8.4
shiboken6==6.4.2
shortuuid==1.0.9
shotdetect-scenedetect-lgss==0.0.3
simplejson==3.18.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.4
stack-data==0.6.2
stanza==1.4.2
stevedore==4.1.1
subword-nmt==0.3.8
sympy==1.11.1
tabulate==0.8.10
taming-transformers-rom1504==0.0.6
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow-estimator==2.10.0
termcolor==2.1.1
terminado==0.17.1
terminaltables==3.1.10
text-unidecode==1.3
text2sql-lgesql==1.2.0
tf-slim==1.1.0
thinc==8.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.5.4
tinycss2==1.2.1
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1+cu113
torchmetrics==0.10.0
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers==4.25.1
typeguard==2.13.3
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.3.0
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.12
visualdl==2.4.1
wandb==0.13.3
wasabi==0.10.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
xlrd==2.0.1
xtcocotools==1.12
xxhash==3.0.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
zhconv==1.4.3
zipp==3.11.0
